feat(#133): rename auth_showcase → keycard_showcase + add signing demo
- Rename auth_showcase-{core,ui} → keycard_showcase-{core,ui}; update
  all manifest names, cmake targets, plugin class, and QML callModule refs
- Add hashMessage() Q_INVOKABLE to keycard_showcase plugin (SHA-256 via
  QCryptographicHash) so QML can hash the message before calling requestSign
- keycard_showcase-ui Main.qml: add Message Signing section with editable
  input (pre-filled "hello world"), Sign with Keycard button, pending
  indicator, and Schnorr signature result display
- PinEntryScreen.qml: checkPendingRequests polls both getPendingAuths and
  getPendingSignRequests; approveRequest/declineRequest dispatch to
  approveSign/declineSign for sign requests; request view shows payload
  hash + scheme for sign requests; Approve button label changes to Sign;
  activity log shows "Signed — return to {caller}" on completion
