--- auth_showcase-core/modules/auth_showcase/manifest.json
@@ -1,16 +1,0 @@
-{
-  "author": "alisher",
-  "category": "demo",
-  "dependencies": ["keycard"],
-  "description": "Auth showcase module - demonstrates Keycard integration",
-  "icon": "",
-  "main": {
-    "linux-amd64": "auth_showcase_plugin.so",
-    "darwin-arm64": "auth_showcase_plugin.dylib"
-  },
-  "manifestVersion": "0.1.0",
-  "name": "auth_showcase",
-  "capabilities": ["auth-showcase-demo"],
-  "type": "core",
-  "version": "1.0.0"
-}
