--- auth_showcase-core/src/plugin.cpp
@@ -1,20 +1,0 @@
-#include "plugin.h"
-#include <QJsonDocument>
-#include <QJsonObject>
-
-AuthShowcasePlugin::AuthShowcasePlugin(QObject* parent)
-    : QObject(parent)
-{
-}
-
-void AuthShowcasePlugin::initLogos(LogosAPI* api)
-{
-    logosAPI = api;
-}
-
-QString AuthShowcasePlugin::initialize()
-{
-    QJsonObject result;
-    result["initialized"] = true;
-    return QJsonDocument(result).toJson(QJsonDocument::Compact);
-}
