--- auth_showcase-core/src/plugin_metadata.json
@@ -1,10 +1,0 @@
-{
-  "author": "alisher",
-  "category": "demo",
-  "dependencies": ["keycard"],
-  "description": "Auth showcase module - demonstrates Keycard integration",
-  "main": "auth_showcase_plugin",
-  "name": "auth_showcase",
-  "type": "core",
-  "version": "1.0.0"
-}
