--- auth_showcase-ui/plugins/auth_showcase-ui/manifest.json
@@ -1,14 +1,0 @@
-{
-  "author": "alisher",
-  "capabilities": ["auth-showcase-debug-ui"],
-  "category": "demo",
-  "dependencies": ["keycard", "auth_showcase"],
-  "description": "Auth showcase UI - demonstrates Keycard integration",
-  "icon": "",
-  "main": {},
-  "manifestVersion": "0.2.0",
-  "name": "auth_showcase-ui",
-  "type": "ui_qml",
-  "version": "1.0.0",
-  "view": "Main.qml"
-}
